Cut 4 Part Complete
Begun work on Cut 4 (Panda Swallow), adding a straight line illustration for the mouth with a mixture of mesh and puppet warp to shape it to expressions, blending to the original where necessary.
Fixed the way panda ears are precomped, animating the Ear comp and using that inside Ear Mask before compiling them into the complete comp.